fix campaign launcher api calls
