+++ campaign-launcher/interface/vite.config.ts
@@ -8,11 +8,6 @@ export default defineConfig(() => {
   return {
     server: {
       port: 3000,
-      proxy: {
-        '/api': {
-          target: 'http://localhost:5000',
-        },
-      },
     },
     plugins: [
       nodePolyfills({
