--- reputation-oracle/vercel.json
@@ -1,101 +1,0 @@
-{
-  "version": 2,
-  "buildCommand": "yarn build && yarn migration:run",
-  "outputDirectory": "dist",
-  "functions": {
-    "api/app.ts": {
-      "maxDuration": 300
-    }
-  },
-  "redirects": [{ "source": "/", "destination": "/swagger" }],
-  "rewrites": [
-    {
-      "source": "/swagger",
-      "destination": "api/app.ts"
-    },
-    {
-      "source": "/swagger/(.*)",
-      "destination": "api/app.ts"
-    },
-    {
-      "source": "/api/health",
-      "destination": "api/app.ts"
-    },
-    {
-      "source": "/api/webhook",
-      "destination": "api/app.ts"
-    },
-    {
-      "source": "/api/webhook/(.*)",
-      "destination": "api/app.ts"
-    },
-    {
-      "source": "/api/payout/(.*)",
-      "destination": "api/app.ts"
-    }
-  ],
-  "crons": [
-    {
-      "path": "/api/payout/cron/process-payout",
-      "schedule": "0 08 * * *"
-    },
-    {
-      "path": "/api/payout/cron/finalize-campaign",
-      "schedule": "0 16 * * *"
-    }
-  ],
-  "headers": [
-    {
-      "source": "/api/(.*)",
-      "headers": [
-        {
-          "key": "Access-Control-Allow-Origin",
-          "value": "*"
-        },
-        {
-          "key": "Access-Control-Allow-Methods",
-          "value": "*"
-        },
-        {
-          "key": "Access-Control-Allow-Headers",
-          "value": "*"
-        }
-      ]
-    },
-    {
-      "source": "/api/payout/(.*)",
-      "headers": [
-        {
-          "key": "Access-Control-Allow-Origin",
-          "value": "*"
-        },
-        {
-          "key": "Access-Control-Allow-Methods",
-          "value": "*"
-        },
-        {
-          "key": "Access-Control-Allow-Headers",
-          "value": "*"
-        }
-      ]
-    },
-    {
-      "source": "/api/webhook/(.*)",
-      "headers": [
-        {
-          "key": "Access-Control-Allow-Origin",
-          "value": "*"
-        },
-        {
-          "key": "Access-Control-Allow-Methods",
-          "value": "*"
-        },
-        {
-          "key": "Access-Control-Allow-Headers",
-          "value": "*"
-        }
-      ]
-    }
-  ],
-  "ignoreCommand": "git diff HEAD^ HEAD --quiet -- ."
-}
